+++ packages/backend/src/models/SearchModel/index.ts
@@ -7,7 +7,6 @@ import {
     isDimension,
     isExploreError,
     NotExistsError,
-    ParameterError,
     SavedChartSearchResult,
     SearchFilters,
     SearchItemType,
@@ -18,14 +17,22 @@ import {
     TableSelectionType,
 } from '@lightdash/common';
 import { Knex } from 'knex';
-import moment from 'moment';
-import { DashboardsTableName } from '../../database/entities/dashboards';
+import {
+    DashboardsTableName,
+    DashboardVersionsTableName,
+} from '../../database/entities/dashboards';
 import {
     CachedExploresTableName,
     ProjectTableName,
 } from '../../database/entities/projects';
 import { SavedChartsTableName } from '../../database/entities/savedCharts';
 import { SpaceTableName } from '../../database/entities/spaces';
+import { UserTableName } from '../../database/entities/users';
+import {
+    filterByCreatedAt,
+    filterByCreatedByUuid,
+    shouldSearchForType,
+} from './utils/filters';
 import { getFullTextSearchRankCalcSql } from './utils/fullTextSearch';
 
 type ModelDependencies = {
@@ -39,72 +46,12 @@ export class SearchModel {
         this.database = deps.database;
     }
 
-    private static shouldSearchForType(
-        entityType: SearchItemType,
-        queryTypeFilter?: string,
-    ) {
-        // if there is no filter or if the filter is the same as the entityType
-        return !queryTypeFilter || queryTypeFilter === entityType;
-    }
-
-    private static filterByCreatedAt<T extends {}, R>(
-        tableName: string,
-        query: Knex.QueryBuilder<T, R>,
-        filters: SearchFilters = {},
-    ) {
-        const { fromDate, toDate } = filters;
-        const fromDateObj = fromDate ? moment(fromDate).utc() : undefined;
-        const toDateObj = toDate ? moment(toDate).utc() : undefined;
-        const now = moment();
-
-        if (fromDateObj?.isAfter(toDateObj)) {
-            throw new ParameterError('fromDate cannot be after toDate');
-        }
-
-        if (fromDateObj) {
-            if (!fromDateObj.isValid()) {
-                throw new ParameterError('fromDate is not valid');
-            }
-
-            if (fromDateObj.isAfter(now)) {
-                throw new ParameterError('fromDate cannot be in the future');
-            }
-
-            query.whereRaw(
-                `Date(${tableName}.created_at) >= ?`,
-                fromDateObj.startOf('day').toDate(),
-            );
-        }
-
-        if (toDateObj) {
-            if (!toDateObj.isValid()) {
-                throw new ParameterError('toDate is not valid');
-            }
-
-            if (toDateObj.isAfter(now)) {
-                throw new ParameterError('toDate cannot be in the future');
-            }
-
-            query.whereRaw(
-                `Date(${tableName}.created_at) <= ?`,
-                toDateObj.endOf('day').toDate(),
-            );
-        }
-
-        return query;
-    }
-
     private async searchSpaces(
         projectUuid: string,
         query: string,
         filters?: SearchFilters,
     ): Promise<SpaceSearchResult[]> {
-        if (
-            !SearchModel.shouldSearchForType(
-                SearchItemType.SPACE,
-                filters?.type,
-            )
-        ) {
+        if (!shouldSearchForType(SearchItemType.SPACE, filters?.type)) {
             return [];
         }
 
@@ -116,7 +63,7 @@ export class SearchModel {
                 query,
             );
 
-        const baseSubquery = this.database(SpaceTableName)
+        let subquery = this.database(SpaceTableName)
             .innerJoin(
                 ProjectTableName,
                 `${ProjectTableName}.project_id`,
@@ -126,9 +73,18 @@ export class SearchModel {
             .where('projects.project_uuid', projectUuid)
             .orderBy(searchRankColumnName, 'desc');
 
-        const subquery = SearchModel.filterByCreatedAt(
-            SpaceTableName,
-            baseSubquery,
+        subquery = filterByCreatedAt(SpaceTableName, subquery, filters);
+        subquery = filterByCreatedByUuid(
+            subquery,
+            {
+                join: {
+                    joinTableName: UserTableName,
+                    joinTableIdColumnName: 'user_id',
+                    joinTableUserUuidColumnName: 'user_uuid',
+                    tableIdColumnName: 'created_by_user_id',
+                },
+                tableName: SpaceTableName,
+            },
             filters,
         );
 
@@ -144,12 +100,7 @@ export class SearchModel {
         query: string,
         filters?: SearchFilters,
     ): Promise<DashboardSearchResult[]> {
-        if (
-            !SearchModel.shouldSearchForType(
-                SearchItemType.DASHBOARD,
-                filters?.type,
-            )
-        ) {
+        if (!shouldSearchForType(SearchItemType.DASHBOARD, filters?.type)) {
             return [];
         }
 
@@ -161,7 +112,7 @@ export class SearchModel {
                 query,
             );
 
-        const baseSubquery = this.database(DashboardsTableName)
+        let subquery = this.database(DashboardsTableName)
             .leftJoin(
                 SpaceTableName,
                 `${DashboardsTableName}.space_id`,
@@ -182,9 +133,19 @@ export class SearchModel {
             .where(`${ProjectTableName}.project_uuid`, projectUuid)
             .orderBy(searchRankColumnName, 'desc');
 
-        const subquery = SearchModel.filterByCreatedAt(
-            DashboardsTableName,
-            baseSubquery,
+        subquery = filterByCreatedAt(DashboardsTableName, subquery, filters);
+        subquery = filterByCreatedByUuid(
+            subquery,
+            {
+                join: {
+                    isVersioned: true,
+                    joinTableName: DashboardVersionsTableName,
+                    joinTableIdColumnName: 'dashboard_id',
+                    joinTableUserUuidColumnName: 'updated_by_user_uuid',
+                    tableIdColumnName: 'dashboard_id',
+                },
+                tableName: DashboardsTableName,
+            },
             filters,
         );
 
@@ -227,12 +188,7 @@ export class SearchModel {
         query: string,
         filters?: SearchFilters,
     ): Promise<SavedChartSearchResult[]> {
-        if (
-            !SearchModel.shouldSearchForType(
-                SearchItemType.CHART,
-                filters?.type,
-            )
-        ) {
+        if (!shouldSearchForType(SearchItemType.CHART, filters?.type)) {
             return [];
         }
 
@@ -245,7 +201,7 @@ export class SearchModel {
             );
 
         // Needs to be a subquery to be able to use the search rank column to filter out 0 rank results
-        const baseSubquery = this.database(SavedChartsTableName)
+        let subquery = this.database(SavedChartsTableName)
             .leftJoin(
                 SpaceTableName,
                 `${SavedChartsTableName}.space_id`,
@@ -270,15 +226,19 @@ export class SearchModel {
             .where(`${ProjectTableName}.project_uuid`, projectUuid)
             .orderBy(searchRankColumnName, 'desc');
 
-        const subQuery = SearchModel.filterByCreatedAt(
-            SavedChartsTableName,
-            baseSubquery,
+        subquery = filterByCreatedAt(SavedChartsTableName, subquery, filters);
+        subquery = filterByCreatedByUuid(
+            subquery,
+            {
+                tableName: SavedChartsTableName,
+                tableUserUuidColumnName: 'last_version_updated_by_user_uuid',
+            },
             filters,
         );
 
         const savedCharts = await this.database(SavedChartsTableName)
             .select()
-            .from(subQuery.as('saved_charts_with_rank'))
+            .from(subquery.as('saved_charts_with_rank'))
             .where(searchRankColumnName, '>', 0)
             .limit(10);
 
@@ -356,12 +316,12 @@ export class SearchModel {
         explores: Explore[],
         filters?: SearchFilters,
     ) {
-        const shouldSearchForTables = SearchModel.shouldSearchForType(
+        const shouldSearchForTables = shouldSearchForType(
             SearchItemType.TABLE,
             filters?.type,
         );
 
-        const shouldSearchForFields = SearchModel.shouldSearchForType(
+        const shouldSearchForFields = shouldSearchForType(
             SearchItemType.FIELD,
             filters?.type,
         );
@@ -483,9 +443,7 @@ export class SearchModel {
         query: string,
         filters?: SearchFilters,
     ) {
-        if (
-            !SearchModel.shouldSearchForType(SearchItemType.PAGE, filters?.type)
-        ) {
+        if (!shouldSearchForType(SearchItemType.PAGE, filters?.type)) {
             return [];
         }