+++ packages/common/src/authorization/projectMemberAbility.ts
@@ -30,6 +30,9 @@ export const projectMemberAbilities: Record<
         can('manage', 'ExportCsv', {
             projectUuid: member.projectUuid,
         });
+        can('view', 'DashboardComments', {
+            projectUuid: member.projectUuid,
+        });
     },
     interactive_viewer(member, { can }) {
         projectMemberAbilities.viewer(member, { can });
@@ -43,6 +46,9 @@ export const projectMemberAbilities: Record<
             projectUuid: member.projectUuid,
         });
         can('create', 'ScheduledDeliveries', {
+            projectUuid: member.projectUuid,
+        });
+        can('create', 'DashboardComments', {
             projectUuid: member.projectUuid,
         });
     },
@@ -65,6 +71,9 @@ export const projectMemberAbilities: Record<
             projectUuid: member.projectUuid,
         });
         can('manage', 'ScheduledDeliveries', {
+            projectUuid: member.projectUuid,
+        });
+        can('manage', 'DashboardComments', {
             projectUuid: member.projectUuid,
         });
     },