+++ packages/frontend/src/components/DashboardTiles/DashboardLoomTile.tsx
@@ -1,5 +1,7 @@
 import { DashboardLoomTile } from '@lightdash/common';
-import React, { FC } from 'react';
+import React, { FC, useState } from 'react';
+import { DashboardTileComments } from '../../features/comments';
+import { useDashboardContext } from '../../providers/DashboardProvider';
 import TileBase from './TileBase/index';
 import { getLoomId } from './TileForms/LoomTileForm';
 
@@ -9,13 +11,33 @@ type Props = Pick<
 > & { tile: DashboardLoomTile };
 
 const LoomTile: FC<Props> = (props) => {
+    const showComments = useDashboardContext(
+        (c) =>
+            c.dashboardCommentsCheck?.userCanViewDashboardComments &&
+            c.dashboardCommentsCheck?.isDashboardTileCommentsFeatureEnabled,
+    );
+    const [isCommentsMenuOpen, setIsCommentsMenuOpen] = useState(false);
     const {
         tile: {
             properties: { title, url },
         },
     } = props;
     return (
-        <TileBase title={title} {...props}>
+        <TileBase
+            title={title}
+            lockHeaderVisibility={isCommentsMenuOpen}
+            extraHeaderElement={
+                !!showComments && (
+                    <DashboardTileComments
+                        opened={isCommentsMenuOpen}
+                        onOpen={() => setIsCommentsMenuOpen(true)}
+                        onClose={() => setIsCommentsMenuOpen(false)}
+                        dashboardTileUuid={props.tile.uuid}
+                    />
+                )
+            }
+            {...props}
+        >
             <iframe
                 title={title}
                 className="non-draggable"