+++ packages/frontend/src/features/comments/components/DashboardTileComments.tsx
@@ -1,0 +1,100 @@
+import { ActionIcon, Popover, PopoverProps, Stack, Text } from '@mantine/core';
+import { IconMessage } from '@tabler/icons-react';
+import { FC } from 'react';
+import MantineIcon from '../../../components/common/MantineIcon';
+import { useApp } from '../../../providers/AppProvider';
+import { useDashboardContext } from '../../../providers/DashboardProvider';
+import { useCreateComment } from '../hooks/useComments';
+import { CommentForm } from './CommentForm';
+import { DashboardCommentAndReplies } from './DashboardCommentAndReplies';
+
+type Props = {
+    dashboardTileUuid: string;
+};
+
+export const DashboardTileComments: FC<
+    Props & Pick<PopoverProps, 'opened' | 'onClose' | 'onOpen'>
+> = ({ dashboardTileUuid, opened, onClose, onOpen }) => {
+    const { user } = useApp();
+    const projectUuid = useDashboardContext((c) => c.projectUuid);
+    const dashboardUuid = useDashboardContext((c) => c.dashboard?.uuid);
+    const userCanManageDashboardComments = useDashboardContext(
+        (c) => c.dashboardCommentsCheck?.userCanManageDashboardComments,
+    );
+    const comments = useDashboardContext(
+        (c) => c.dashboardComments && c.dashboardComments[dashboardTileUuid],
+    );
+    const { mutateAsync, isLoading } = useCreateComment();
+
+    if (!projectUuid || !dashboardUuid) {
+        return null;
+    }
+
+    return (
+        <Popover
+            withArrow
+            withinPortal
+            shadow="md"
+            position="bottom-end"
+            offset={4}
+            arrowOffset={10}
+            opened={opened}
+            onOpen={() => {
+                onOpen?.();
+            }}
+            onClose={() => {
+                onClose?.();
+            }}
+        >
+            <Popover.Dropdown miw={400}>
+                <Stack
+                    spacing="xs"
+                    sx={{
+                        maxHeight: 300,
+                        overflowY: 'auto',
+                    }}
+                >
+                    {comments?.map((comment) => (
+                        <DashboardCommentAndReplies
+                            key={comment.commentId}
+                            comment={comment}
+                            projectUuid={projectUuid}
+                            dashboardUuid={dashboardUuid}
+                            dashboardTileUuid={dashboardTileUuid}
+                        />
+                    ))}
+                    {!userCanManageDashboardComments &&
+                        (!comments ||
+                            (comments.length === 0 && (
+                                <Text fz="xs">No comments yet</Text>
+                            )))}
+                </Stack>
+                {userCanManageDashboardComments && (
+                    <CommentForm
+                        userName={
+                            user.data?.firstName + ' ' + user.data?.lastName
+                        }
+                        onSubmit={(text: string) =>
+                            mutateAsync({
+                                projectUuid,
+                                dashboardUuid,
+                                dashboardTileUuid,
+                                text,
+                            })
+                        }
+                        isSubmitting={isLoading}
+                    />
+                )}
+            </Popover.Dropdown>
+
+            <Popover.Target>
+                <ActionIcon
+                    size="sm"
+                    onClick={() => (opened ? onClose?.() : onOpen?.())}
+                >
+                    <MantineIcon icon={IconMessage} />
+                </ActionIcon>
+            </Popover.Target>
+        </Popover>
+    );
+};